+++ kube_scheduler/datadog_checks/kube_scheduler/kube_scheduler.py
@@ -13,6 +13,8 @@ from datadog_checks.base.checks.openmetrics import OpenMetricsBaseCheck
 from datadog_checks.base.config import is_affirmative
 from datadog_checks.base.utils.http import RequestsWrapper
 
+from .sli_metrics import SliMetricsScraperMixin
+
 DEFAULT_COUNTERS = {
     # Number of HTTP requests, partitioned by status code, method, and host.
     'rest_client_requests_total': 'client.http.requests',
@@ -131,7 +133,7 @@ IGNORE_METRICS = [
 ]
 
 
-class KubeSchedulerCheck(KubeLeaderElectionMixin, OpenMetricsBaseCheck):
+class KubeSchedulerCheck(KubeLeaderElectionMixin, SliMetricsScraperMixin, OpenMetricsBaseCheck):
     DEFAULT_METRIC_LIMIT = 0
 
     KUBE_SCHEDULER_NAMESPACE = "kube_scheduler"
@@ -183,9 +185,15 @@ class KubeSchedulerCheck(KubeLeaderElectionMixin, OpenMetricsBaseCheck):
 
                 instance['health_url'] = url
 
+        inst = instances[0] if instances else None
+        slis_instance = self.create_sli_prometheus_instance(inst)
+        self.slis_scraper_config = self.get_scraper_config(slis_instance)
+        self.detect_sli_endpoint(self.get_http_handler(self.slis_scraper_config), slis_instance.get('prometheus_url'))
+
     def check(self, instance):
         # Get the configuration for this specific instance
         scraper_config = self.get_scraper_config(instance)
+
         # Set up metric_transformers
         transformers = {}
         for metric_from, metric_to in TRANSFORM_VALUE_HISTOGRAMS.items():
@@ -202,6 +210,10 @@ class KubeSchedulerCheck(KubeLeaderElectionMixin, OpenMetricsBaseCheck):
             self.check_election_status(leader_config)
 
         self._perform_service_check(instance)
+
+        if self._slis_available:
+            self.log.debug('processing kube scheduler sli metrics')
+            self.process(self.slis_scraper_config)
 
     def _perform_service_check(self, instance):
         url = instance.get('health_url')